Compress or move methods to declutter context.py
